Export mock data to own files, Add workflow.d.ts and Add workflow functionality
