+++ frontend/app/utils/project.d.ts
@@ -9,6 +9,5 @@ export interface Project {
   targetPlatform: ProjectTargetPlatform[] | ProjectTargetPlatform | string | string[]
   created_at: string
   updated_at: string
-  // Optional icon data URL (base64) or remote URL
   icon?: string | null
 }
